+++ CHANGELOG.md
@@ -68,6 +68,8 @@
 - Bug-073127 - The banner is not properly aligned even if the admin added the banner according to the preferred dimensions.
 - Bug-073166 - Order net amount in sales reports displaying negative for the seller dashboard section
 - Bug-069186 - Android, iOS >> warning appears in the email received for the order details
+- Bug-067979 - Android, iOS >> Product warranty is not correct on the product details page on mobile devices
+
 # RV-10.0.0.20221123
 
 ## Bug Fixes: